refactor: DTO 모델/데이터파일로 리팩토
