+++ frontend/app/src/main/java/com/markoala/tomoandroid/data/model/ApiResponse.kt
@@ -1,0 +1,6 @@
+package com.markoala.tomoandroid.data.model
+
+data class PostResponse(
+    val success: Boolean,
+    val message: String
+)
