+++ frontend/app/src/main/java/com/markoala/tomoandroid/data/model/MeetingData.kt
@@ -1,0 +1,9 @@
+package com.markoala.tomoandroid.data.model
+
+
+data class MeetingSummary(
+    val title: String,
+    val location: String?,
+    val time: String?,
+    val peopleCounts: Int = 1
+)
