+++ frontend/app/src/main/java/com/markoala/tomoandroid/data/model/UserData.kt
@@ -1,12 +1,16 @@
 package com.markoala.tomoandroid.data.model
 
+import java.time.LocalDate
+
 data class UserData(
     val uuid: String,
     val email: String,
     val username: String
 )
 
-data class PostResponse(
-    val success: Boolean,
-    val message: String
+data class FriendProfile(
+    val name: String,
+    val email: String,
+    val friendSince: LocalDate,
+    val intimacy: Int
 )
