+++ frontend/app/src/main/java/com/markoala/tomoandroid/ui/main/HomeScreen.kt
@@ -25,10 +25,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.markoala.tomoandroid.data.model.MeetingSummary
 import com.markoala.tomoandroid.ui.components.CustomText
 import com.markoala.tomoandroid.ui.components.CustomTextType
 import com.markoala.tomoandroid.ui.components.home.MeetingCard
-import com.markoala.tomoandroid.ui.components.home.MeetingSummary
+
 import com.markoala.tomoandroid.ui.theme.CustomColor
 
 private val sampleMeetings = listOf(
